modules: split nextshow
